sandbox: add slug
