+++ packages/ui/src/components/Forms/SlugControl/index.sass
@@ -1,1 +1,1 @@
-@import 'SlugControl'
+@import 'slugInput'
