+++ packages/ui/src/components/Forms/SlugControl/index.ts
@@ -1,1 +1,1 @@
-export * from './SlugControl'
+export * from './SlugInput'
