+++ packages/ui/src/components/Forms/SlugControl/slugInput.sass
@@ -1,4 +1,4 @@
-.#{$cui-conf-globalPrefix}slugControl
+.#{$cui-conf-globalPrefix}slugInput
 	display: flex
 	align-items: stretch
